backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
pydantic-settings
email-validator
bcrypt
google-auth
requests
httpx
python-multipart
python-dotenv
firebase-admin
google-cloud-aiplatform
google-cloud-storage
google-cloud-firestore
google-cloud-discoveryengine
langchain>=1.1.0
langchain-core>=0.3.0
langchain-google-vertexai
langgraph>=0.2.0
google-api-python-client
pypdf
reportlab
